Address review: move helpers to conftest, use parametrize for alarm tests
