+++ tests/components/touchline_sl/conftest.py
@@ -2,7 +2,7 @@
 
 from collections.abc import Generator
 from typing import NamedTuple
-from unittest.mock import AsyncMock, patch
+from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
@@ -17,6 +17,39 @@ class FakeModule(NamedTuple):
 
     name: str
     id: str
+
+
+def make_mock_zone(
+    zone_id: int = 1, name: str = "Zone 1", alarm: str | None = None
+) -> MagicMock:
+    """Return a mock Zone with configurable alarm state."""
+    zone = MagicMock()
+    zone.id = zone_id
+    zone.name = name
+    zone.temperature = 21.5
+    zone.target_temperature = 22.0
+    zone.humidity = 45
+    zone.mode = "constantTemp"
+    zone.algorithm = "heating"
+    zone.relay_on = False
+    zone.alarm = alarm
+    zone.schedule = None
+    zone.enabled = True
+    zone.signal_strength = 100
+    zone.battery_level = None
+    return zone
+
+
+def make_mock_module(zones: list) -> MagicMock:
+    """Return a mock module with the given zones."""
+    module = MagicMock()
+    module.id = "deadbeef"
+    module.name = "Foobar"
+    module.type = "SL"
+    module.version = "1.0"
+    module.zones = AsyncMock(return_value=zones)
+    module.schedules = AsyncMock(return_value=[])
+    return module
 
 
 @pytest.fixture
@@ -44,6 +77,18 @@ def mock_touchlinesl_client() -> Generator[AsyncMock]:
         client = mock_client.return_value
         client.user_id.return_value = 12345
         client.modules.return_value = [FakeModule(name="Foobar", id="deadbeef")]
+        yield client
+
+
+@pytest.fixture
+def mock_touchlinesl_full_client() -> Generator[MagicMock]:
+    """Mock a pytouchlinesl client with full module/zone support."""
+    with patch(
+        "homeassistant.components.touchline_sl.TouchlineSL",
+        autospec=True,
+    ) as mock_client:
+        client = mock_client.return_value
+        client.user_id = AsyncMock(return_value=12345)
         yield client
